requirements.txt
opencv-python
ultralytics
numpy
pillow
openai